Refactoring for Conditional Polymorphism and pull up members 1. Handled the conditional polymorphism in "CommonDebug" class by creating seperate classes for all four command types - Config, Exp, Party, Player and moved the parseDebugExp method to the appropriate class. 2. Created a ServerCommonPlugin class to add another layer between velocity and Bungee parties plugin as getServerId method was same in both the classes hence, could be moved up. 3. Extracted a new class out of "CommonListener" named "CommonPartyListener" to divide the responsibility of a single class. All the methods related to Party are shifted to this newly created class.
Note - All the changes made above have been tested with a successful "mvn clean install" and not causing any issue in the existing code.
