+++ common/src/main/java/com/alessiodp/parties/common/commands/sub/CommandStart.java
@@ -1,0 +1,6 @@
+package com.alessiodp.parties.common.commands.sub;
+
+public interface CommandStart {
+
+    public void commandStart();
+}
